+++ src/main/kotlin/net/internetisalie/lunar/lang/navigation/LuaGotoClassContributor.kt
@@ -13,7 +13,6 @@ import com.intellij.util.Processor
 import com.intellij.util.indexing.FindSymbolParameters
 import com.intellij.util.indexing.IdFilter
 import net.internetisalie.lunar.lang.LuaLanguage
-import net.internetisalie.lunar.lang.indexing.LuaAliasIndex
 import net.internetisalie.lunar.lang.indexing.LuaClassNameIndex
 import net.internetisalie.lunar.lang.psi.LuaLocalVarDecl
 import javax.swing.Icon
@@ -25,9 +24,8 @@ import javax.swing.Icon
 class LuaGotoClassContributor : GotoClassContributor, ChooseByNameContributorEx {
 
     override fun processNames(processor: Processor<in String>, scope: GlobalSearchScope, filter: IdFilter?) {
-        val index = StubIndex.getInstance()
-        index.processAllKeys(LuaClassNameIndex.KEY, processor, scope, filter)
-        index.processAllKeys(LuaAliasIndex.KEY, processor, scope, filter)
+        StubIndex.getInstance().processAllKeys(LuaClassNameIndex.KEY, processor, scope, filter)
+        LuaAliasNavigation.processNames(processor, scope, filter)
     }
 
     override fun processElementsWithName(
@@ -41,7 +39,7 @@ class LuaGotoClassContributor : GotoClassContributor, ChooseByNameContributorEx 
         val aliasIcon = AllIcons.Nodes.Type
         val emitClasses = emit(LuaClassNameIndex.KEY, name, project, scope, classIcon, processor)
         if (!emitClasses) return
-        emit(LuaAliasIndex.KEY, name, project, scope, aliasIcon, processor)
+        LuaAliasNavigation.processElements(name, project, scope, aliasIcon, processor)
     }
 
     private fun emit(